+++ smac/optimizer/configuration_chooser/boing_chooser.py
@@ -14,7 +14,10 @@ from smac.epm.utils import get_types
 from smac.optimizer.acquisition import EI, TS, AbstractAcquisitionFunction
 from smac.optimizer.acquisition.maximizer import AcquisitionFunctionMaximizer
 from smac.optimizer.configuration_chooser.epm_chooser import EPMChooser
-from smac.optimizer.configuration_chooser.random_chooser import ChooserNoCoolDown, RandomChooser
+from smac.optimizer.configuration_chooser.random_chooser import (
+    ChooserNoCoolDown,
+    RandomChooser,
+)
 from smac.optimizer.subspaces.boing_subspace import BOinGSubspace
 from smac.optimizer.subspaces.turbo_subspace import TuRBOSubSpace
 from smac.runhistory.runhistory import RunHistory